+++ src/presenter/trip-presenter.js
@@ -1,6 +1,6 @@
 import TripPointView from '../view/trip-point.js';
 import TripPointsList from '../view/trip-point-list.js';
-import NewPointView from '../view/new-point-view.js';
+import FormEditTemplate from '../view/form-edit-view.js';
 import SortView from '../view/sort-view.js';
 import { render } from '../render.js';
 
@@ -15,7 +15,7 @@ export default class TripPresenter {
   init() {
     render(this.sortComponent, this.tripContainer);
     render(this.tripComponent, this.tripContainer);
-    render (new NewPointView(), this.tripComponent.getElement());
+    render (new FormEditTemplate(), this.tripComponent.getElement());
 
     for (let i = 0; i < 3; i++) {
       render(new TripPointView(), this.tripComponent.getElement());
